+++ backend/main.py
@@ -1,27 +1,38 @@
 from fastapi import FastAPI, Response, status
 from pydantic import BaseModel
-from backend.handle_tokens import decode_jwt
-from backend.generate_keys import read_public_key
+from backend.handle_tokens import decode_jwt, encode_jwt
+from backend.github_auth import request_access_token
+from settings import PRIV_KEY_PATH, PUB_KEY_PATH
+
 
 app = FastAPI()
 
 
-class Item(BaseModel):
-    jwt: dict
+class Token(BaseModel):
+    encoded_jwt: str
 
 
 @app.get("/")
 def read_root():
     return "Alrighty then"
 
-@app.get("/token", status_code=200)
-def create_item(jwt: Item, response: Response):
-    decoded = decode_jwt(jwt, read_public_key("test.pub"))
-    if decoded is None:
-        response.status_code = status.HTTP_403_FORBIDDEN
-        return response.status_code
-    else:
-        response.status_code = status.HTTP_200_OK
-        return {"status": response.status_code,
-                "payload": decoded}
 
+@app.get("/auth/{code}")
+def return_jwt(code):
+    at = request_access_token(code)
+    if at == "Invalid Request":
+        return at, status.HTTP_403_FORBIDDEN
+    else:
+        print(PRIV_KEY_PATH)
+        encoded = encode_jwt(at, PRIV_KEY_PATH)
+        return encoded
+
+
+@app.post("/teams")
+def teams(jwt: Token, user):
+    # Decode and verify the jwt
+    try:
+        decoded = decode_jwt(jwt.encoded_jwt, PUB_KEY_PATH)
+        return decoded, user
+    except ValueError:
+        return status.HTTP_403_FORBIDDEN
