+++ requirements.txt
@@ -1,0 +1,21 @@
+certifi==2020.11.8
+cffi==1.14.3
+chardet==3.0.4
+click==7.1.2
+cryptography==3.2.1
+Deprecated==1.2.10
+fastapi==0.61.2
+h11==0.11.0
+idna==2.10
+psycopg2==2.8.6
+pycparser==2.20
+pydantic==1.7.2
+PyGithub==1.53
+PyJWT==1.7.1
+python-dotenv==0.15.0
+requests==2.25.0
+six==1.15.0
+starlette==0.13.6
+urllib3==1.26.2
+uvicorn==0.12.2
+wrapt==1.12.1
